requirements.txt
crawl4ai
python-dotenv==1.0.1
duckduckgo-search-api==0.1.5
qdrant-client==1.12.0
chromadb==0.5.15
semantic-text-splitter==0.18.0